moves App outsides of components folder to be a parent component
